Keeping methods at their location does not work reliably.

+++ src/main/kotlin/andrasferenczi/action/utils/constructorDeleteProcess.kt
@@ -5,17 +5,49 @@ import com.intellij.openapi.project.Project
 import com.intellij.openapi.ui.Messages
 import com.jetbrains.lang.dart.psi.DartClass
 
-fun createConstructorDeleteCallWithUserPrompt(
+// Try to put the constructor after the last
+fun tryGetFirstNamedConstructorStart(
+    dartClass: DartClass
+): Int? {
+    return dartClass.extractMethodConstructorInfos()
+        .firstOrNull { it.constructorType == DartConstructorType.NamedParameterOnly }
+        ?.psi
+        ?.textRange
+        ?.startOffset
+}
+
+fun tryGetLastConstructorEnd(
+    dartClass: DartClass
+): Int? {
+    return dartClass.extractMethodConstructorInfos()
+        .lastOrNull()
+        ?.psi
+        ?.textRange
+        ?.endOffset
+}
+
+fun createConstructorPreCleanProcess(
     project: Project,
     dartClass: DartClass
-): (() -> Unit)? {
+): CodeGenerationPreCleanProcess {
     val additionalCalls: MutableList<() -> Unit> = ArrayList()
 
     val existingInfos = dartClass.extractMethodConstructorInfos()
+    var caretOffsetAfterDelete: Int? = null
     if (existingInfos.isNotEmpty()) {
         val (namedConstructorInfos, otherConstructorInfos) = existingInfos.partition { it.constructorType == DartConstructorType.NamedParameterOnly }
 
-        additionalCalls += { namedConstructorInfos.deleteAllPsiElements() }
+        if (namedConstructorInfos.isNotEmpty()) {
+            // Can only put it to it's start safely
+            caretOffsetAfterDelete = namedConstructorInfos
+                .first()
+                .psi
+                .textRange
+                .startOffset
+
+            additionalCalls += { namedConstructorInfos.deleteAllPsiElements() }
+        }
+
 
         if (otherConstructorInfos.isNotEmpty()) {
             val shouldDeleteInt = Messages.showOkCancelDialog(
@@ -30,10 +62,17 @@ fun createConstructorDeleteCallWithUserPrompt(
             val wantsToDelete = shouldDeleteInt == Messages.OK
 
             if (wantsToDelete) {
+                // Would have to calculate the spacing and the formatting and lengths of the removed constructor
+                // Let's stay safe
+                caretOffsetAfterDelete = null
+
                 additionalCalls += { existingInfos.deleteAllPsiElements() }
             }
         }
     }
 
-    return additionalCalls.mergeCalls()
+    return CodeGenerationPreCleanProcess(
+        deleteCall = additionalCalls.mergeCalls(),
+        caretOffsetAfterDelete = caretOffsetAfterDelete
+    )
 }
